feat: customer register form
