+++ package.json
@@ -28,6 +28,7 @@
         "react": "^19.0.0",
         "react-dom": "^19.0.0",
         "react-icons": "^5.5.0",
+        "react-images-uploading": "^3.1.7",
         "react-router-dom": "^7.2.0",
         "react-slick": "^0.30.3",
         "react-toastify": "^11.0.5",
