--- src/components/button.jsx
@@ -1,48 +1,0 @@
-import { cva } from 'class-variance-authority'
-import * as React from 'react'
-
-import { cn } from '~/helpers/cn'
-
-const buttonVariants = cva(
-    'inert inline-flex items-center justify-center whitespace-nowrap rounded-lg text-base font-semibold ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50',
-    {
-        variants: {
-            variant: {
-                default: 'bg-primary text-white hover:bg-primary/90 rounded-full',
-                destructive:
-                    'bg-destructive text-destructive-foreground hover:bg-destructive/90',
-                outline: 'border border-white text-white rounded-full',
-                secondary:
-                    'bg-secondary text-secondaryText hover:bg-secondary/80 rounded-full',
-                ghost: 'hover:bg-accent hover:text-accent-foreground',
-                link: 'text-primary underline-offset-4 hover:underline',
-            },
-            size: {
-                default: 'h-10 px-5 py-[9px]',
-                sm: 'h-4 px-3 py-2',
-                lg: 'h-11 rounded-md px-8',
-                icon: 'h-10 w-10',
-            },
-        },
-        defaultVariants: {
-            variant: 'default',
-            size: 'default',
-        },
-    },
-)
-
-const Button = React.forwardRef(
-    ({ className, variant, size, asChild = false, ...props }, ref) => {
-        const Comp = 'button'
-        return (
-            <Comp
-                className={cn(buttonVariants({ variant, size, className }))}
-                ref={ref}
-                {...props}
-            />
-        )
-    },
-)
-Button.displayName = 'Button'
-
-export { Button, buttonVariants }
